refactor: added exceptions related to invalid sessions
i also needed to change `@RequestHeader("Authorization")` because the saved header was "authorization"

+++ src/main/java/ifsul/agileproject/rachadinha/service/impl/UserSessionServiceImpl.java
@@ -1,12 +1,17 @@
-package ifsul.agileproject.rachadinha.session;
+package ifsul.agileproject.rachadinha.service.impl;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import ifsul.agileproject.rachadinha.domain.entity.UserSession;
+import ifsul.agileproject.rachadinha.exceptions.UserNotLoggedInException;
+import ifsul.agileproject.rachadinha.repository.UserSessionRepository;
+import ifsul.agileproject.rachadinha.service.UserSessionService;
+
 @Service
 @AllArgsConstructor
-public class UserSessionService {
+public class UserSessionServiceImpl implements UserSessionService {
 
   private UserSessionRepository userSessionRepository;
 
@@ -15,8 +20,11 @@ public class UserSessionService {
   }
 
   public UserSession getSessionByToken(String token) {
-    return userSessionRepository.findByToken(token)
-      .orElse(null);
+    if (userSessionRepository.findByToken(token) == null) {
+      throw new UserNotLoggedInException();
+    }
+
+    return userSessionRepository.findByToken(token);
   }
 
   public boolean isSessionValid(String token) {
@@ -26,6 +34,9 @@ public class UserSessionService {
 
   @Transactional
   public void invalidateSession(String token) {
+    if (userSessionRepository.findByToken(token) == null) {
+      throw new UserNotLoggedInException();
+    }
     userSessionRepository.deleteByToken(token);
   }
